Refractor the qr code logic
